Host the shell on a wp-admin page instead of a custom frontend URL
The shell previously rendered from a frontend rewrite at /cortext/ via a
custom WP_Rewrite pipeline and its own HTML template. That forced three
separate workarounds just to make the block editor look right:

- mutate $wp_styles['wp-block-library']->src mid-request so
  get_block_editor_settings() resolved the full block CSS instead of
  common.css (frontend + block theme → separate-core-block-assets is true)
- walk WP_Theme_JSON_Resolver per block and inline the styles into
  global-styles so per-block theme.json rules rendered at all
- show_admin_bar(false) to suppress a bar the frontend URL kept adding

Register a top-level wp-admin page (Admin\Screen) and render the React
root directly. In admin context is_admin() is true, which flips
wp_should_load_separate_core_block_assets() to false, so wp-block-library
is registered against the full style.css and global-styles picks up
theme.json rules natively. All three workarounds go away.

Deletes the rewrite rule, query var, template_include/redirect_canonical
filters, the standalone template.php, the MenuLink redirect shim, and
the activation/deactivation rewrite-flush hooks. Fullscreen is handled
by a body.cortext-fullscreen class that hides #wpadminbar,

The router swaps to memory history because the URL is now
wp-admin/admin.php?page=cortext, which browser history can't usefully
route against. A follow-up restores URL persistence via a rewrite rule
for /wp-admin/cortext/... and a ?p= query-param adapter for the
canonical admin URL.
