+++ includes/Admin/Screen.php
@@ -1,0 +1,122 @@
+<?php
+/**
+ * Wp-admin top-level page that hosts the Cortext React shell.
+ *
+ * Lives under `wp-admin/admin.php?page=cortext` so `is_admin()` is true and
+ * the editor inherits admin defaults: `wp_should_load_separate_core_block_assets()`
+ * is false (full block CSS lands on `wp-block-library` + `global-styles`),
+ * the admin bar belongs to wp-admin chrome we hide via body class, and the
+ * REST nonce / user context are already set up.
+ *
+ * @package Cortext
+ */
+
+declare( strict_types=1 );
+
+namespace Cortext\Admin;
+
+final class Screen {
+
+	public const MENU_SLUG      = 'cortext';
+	public const HOOK_SUFFIX    = 'toplevel_page_' . self::MENU_SLUG;
+	private const SCRIPT_HANDLE = 'cortext-shell';
+	private const BODY_CLASS    = 'cortext-fullscreen';
+
+	public function register(): void {
+		add_action( 'admin_menu', array( $this, 'register_menu' ) );
+		add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_assets' ) );
+		add_filter( 'admin_body_class', array( $this, 'add_body_class' ) );
+	}
+
+	public function register_menu(): void {
+		add_menu_page(
+			__( 'Cortext', 'cortext' ),
+			__( 'Cortext', 'cortext' ),
+			'edit_posts',
+			self::MENU_SLUG,
+			array( $this, 'render' ),
+			'dashicons-welcome-write-blog',
+			3
+		);
+	}
+
+	public function render(): void {
+		echo '<div id="cortext-root" class="cortext-root"></div>';
+	}
+
+	public function enqueue_assets( string $hook_suffix ): void {
+		if ( self::HOOK_SUFFIX !== $hook_suffix ) {
+			return;
+		}
+
+		$asset_path = CORTEXT_PATH . 'build/index.asset.php';
+		if ( ! file_exists( $asset_path ) ) {
+			return;
+		}
+
+		/**
+		 * Script asset manifest emitted by @wordpress/scripts.
+		 *
+		 * @var array{dependencies: string[], version: string} $asset
+		 */
+		$asset = require $asset_path;
+
+		wp_enqueue_script(
+			self::SCRIPT_HANDLE,
+			CORTEXT_URL . 'build/index.js',
+			$asset['dependencies'],
+			$asset['version'],
+			true
+		);
+
+		wp_add_inline_script(
+			self::SCRIPT_HANDLE,
+			'window.cortextSettings = ' . wp_json_encode(
+				array(
+					'adminUrl' => admin_url(),
+					'menuSlug' => self::MENU_SLUG,
+				)
+			) . ';',
+			'before'
+		);
+
+		wp_set_script_translations( self::SCRIPT_HANDLE, 'cortext' );
+
+		// Resolve the stylesheets the iframe canvas needs (core block CSS,
+		// theme.json global styles, etc.) and hand them to EditorProvider via
+		// a global. In admin context `wp_should_load_separate_core_block_assets()`
+		// is false, so `wp-block-library` is registered pointing at the full
+		// `style.css` — no src-swap or theme.json loop needed here.
+		$editor_context  = new \WP_Block_Editor_Context( array( 'name' => 'core/edit-post' ) );
+		$editor_settings = get_block_editor_settings( array(), $editor_context );
+
+		wp_add_inline_script(
+			self::SCRIPT_HANDLE,
+			'window.cortextEditorSettings = ' . wp_json_encode( $editor_settings ) . ';',
+			'before'
+		);
+
+		// `wp-edit-blocks` ships `block-editor/content.css` (canvas styles);
+		// `wp-block-editor` adds `block-editor/style.css` (toolbar popover).
+		wp_enqueue_style( 'wp-edit-blocks' );
+		wp_enqueue_style( 'wp-block-editor' );
+
+		$style_path = CORTEXT_PATH . 'build/index.css';
+		if ( file_exists( $style_path ) ) {
+			wp_enqueue_style(
+				self::SCRIPT_HANDLE,
+				CORTEXT_URL . 'build/index.css',
+				array(),
+				$asset['version']
+			);
+		}
+	}
+
+	public function add_body_class( string $classes ): string {
+		$screen = function_exists( 'get_current_screen' ) ? get_current_screen() : null;
+		if ( $screen && self::HOOK_SUFFIX === $screen->id ) {
+			$classes .= ' ' . self::BODY_CLASS;
+		}
+		return $classes;
+	}
+}
