+++ src/index.scss
@@ -5,6 +5,20 @@
 // doesn't register a `wp-interface` style handle, so bundle its stylesheet.
 @use "../node_modules/@wordpress/interface/build-style/style.css";
 
+// Hide wp-admin chrome on the Cortext screen. The React shell is
+// `position: fixed; inset: 0`, so admin scaffolding (toolbar, sidebar,
+// footer) would only matter where it has its own `position: fixed` and a
+// higher z-index — namely #wpadminbar and #adminmenumain. Hide them so
+// they don't paint on top of the canvas; #wpfooter joins the same rule
+// for visual cleanliness while debugging via the React DevTools panel.
+body.cortext-fullscreen {
+	#wpadminbar,
+	#adminmenumain,
+	#wpfooter {
+		display: none;
+	}
+}
+
 .cortext-root {
 	position: fixed;
 	inset: 0;
